+++ packages/adminui-vue/components/controls/UButton.vue
@@ -44,6 +44,10 @@
 export default {
   name: 'UButton',
 
+  inject: {
+    appearance: 'default'
+  },
+
   props: {
     /**
      * Left icon css class